requirements.txt
mistralai
python-dotenv
langchain
langchain-mistralai
langchain-openai
langchain-anthropic
langchain-community
streamlit
langchain-groq